+++ protostar/cairo_testing/cairo_test_execution_environment.py
@@ -16,15 +16,15 @@ from protostar.testing.environments.execution_environment import (
 from protostar.testing.cheatcodes.expect_revert_cheatcode import ExpectRevertContext
 from protostar.testing.hook import Hook
 from protostar.cheatable_starknet.cheatable_cached_state import CheatableCachedState
+from protostar.cheatable_starknet.cheaters.expects import assert_no_expected_calls
+from protostar.protostar_exception import ProtostarException
+from protostar.starknet.cheater import CheaterException
+from protostar.starknet import ReportedException
 
 from .cairo_test_cheatcode_factory import (
     CairoTestCheatcodeFactory,
 )
 from .cairo_test_execution_state import CairoTestExecutionState
-from protostar.starknet.cheaters.expects import assert_no_expected_calls
-from protostar.protostar_exception import ProtostarException
-from protostar.starknet.cheater import CheaterException
-from protostar.starknet import ReportedException
 
 
 class CairoTestExecutionEnvironment(ExecutionEnvironment):
@@ -44,9 +44,10 @@ class CairoTestExecutionEnvironment(ExecutionEnvironment):
                 execution_resources=await self.execute_test_case(function_name)
             )
             try:
-                assert_no_expected_calls(
-                    self.state.starknet.cheatable_state.cheatable_state.expected_contract_calls
-                )
+                if isinstance(self.state.starknet.state.state, CheatableCachedState):
+                    assert_no_expected_calls(
+                        self.state.starknet.state.state.expected_contract_calls
+                    )
             except (ProtostarException, CheaterException) as e:
                 raise ReportedException from e
             return result
