+++ protostar/cheatable_starknet/cheatcodes/call_cairo_cheatcode.py
@@ -3,7 +3,7 @@ from typing import Callable, Optional
 
 from protostar.starknet import CheatcodeException
 from protostar.starknet.cheater import CheaterException
-from protostar.testing.cairo_cheatcodes.cairo_cheatcode import CairoCheatcode
+from protostar.cheatable_starknet.cheatcodes.cairo_cheatcode import CairoCheatcode
 from protostar.starknet.data_transformer import CairoOrPythonData, CairoData
 from protostar.starknet import RawAddress, Address
 
@@ -44,6 +44,7 @@ class CallCairoCheatcode(CairoCheatcode):
                 function_name=function_name,
                 calldata=calldata,
                 cheaters=self.cheaters,
+                unregister_expected_call=self.cheaters.expects.unregister_expected_call,
             )
         except CheaterException as exc:
             raise CheatcodeException(self, exc.message) from exc
