+++ protostar/cheatable_starknet/cheaters/__init__.py
@@ -1,10 +1,12 @@
 from dataclasses import dataclass
 
-from protostar.cheatable_starknet.cheaters.block_info import BlockInfoCairoCheater
-from protostar.cheatable_starknet.cheaters.contracts import ContractsCairoCheater
+from .block_info import BlockInfoCairoCheater
+from .contracts import ContractsCairoCheater
+from .expects import ExpectsCairoCheater
 
 
 @dataclass(frozen=True)
 class CairoCheaters:
     block_info: BlockInfoCairoCheater
     contracts: ContractsCairoCheater
+    expects: ExpectsCairoCheater
